Clean up create-team-logo to remove stale logic
